+++ src/dashboard/dashboard/series/DbSeries.jsx
@@ -1,13 +1,11 @@
 import { useState } from "react";
 import Pagination from "../../../components/pagination/Pagination";
-import { usePerPgaeTvShowQuery} from "../../../redux/features/tv-show/tvShowApi";
+import { usePerPgaeTvShowQuery } from "../../../redux/features/tv-show/tvShowApi";
 import { useDeleteMovieSeriesMutation } from "../../../redux/features/trash/trashApi";
 import toast from "react-hot-toast";
 import { Link } from "react-router-dom";
 import { FiTrash } from "react-icons/fi";
 import { Tooltip, initTE } from "tw-elements";
-
-
 
 const DbSeries = () => {
   initTE({ Tooltip });
@@ -15,10 +13,16 @@ const DbSeries = () => {
   const { data: perPgaeMovie } = usePerPgaeTvShowQuery(currentPage);
   const [deleteMovieSeries] = useDeleteMovieSeriesMutation();
 
-
   const handleDeleteSeason = (id) => {
-    deleteMovieSeries(id);
-    toast.error("Deleted");
+    const shouldDelete = window.confirm(
+      "Are you sure want to delete this movie"
+    );
+    if (shouldDelete) {
+      deleteMovieSeries(id);
+      toast.error("Deleted");
+    } else {
+      console.log("Deletion canceled by user");
+    }
     console.log(id);
   };
 
@@ -64,15 +68,22 @@ const DbSeries = () => {
           <tbody className="divide-y">
             {perPgaeMovie?.data?.data?.map((item, idx) => (
               <tr key={idx} className="odd:bg-gray-50 even:bg-white">
-
                 <td className="px-6 py-4 font-medium flex items-center gap-x-2">
-                  <img  src={item?.poster_image_url}  alt="" className="w-[50px] h-[50px] rounded-full"/>
+                  <img
+                    src={item?.poster_image_url}
+                    alt=""
+                    className="w-[50px] h-[50px] rounded-full"
+                  />
 
-                  <p data-te-toggle="tooltip"  title={ item?.post_title.length > 50 ? item?.post_title : ""}>
-                      {item?.post_title.length > 50 ? `${item?.post_title?.slice(0, 50)}...` : item?.post_title}
-                    </p>
-
-                </td> 
+                  <p
+                    data-te-toggle="tooltip"
+                    title={item?.post_title.length > 50 ? item?.post_title : ""}
+                  >
+                    {item?.post_title.length > 50
+                      ? `${item?.post_title?.slice(0, 50)}...`
+                      : item?.post_title}
+                  </p>
+                </td>
 
                 <td className="px-6 py-4 whitespace-nowrap">
                   {item?.post_type}
